// src/generated/graphql.tsx
export type Maybe<T> = T | null;
// ...
export type Aggregation = {
  __typename?: 'Aggregation';
  /** Attribute code of the aggregation group. */
  attribute_code: Scalars['String']['output'];
  /** The number of options in the aggregation group. */
  count?: Maybe<Scalars['Int']['output']>;
  /** The aggregation display name. */
  label?: Maybe<Scalars['String']['output']>;
  /** Array of options for the aggregation. */
  options?: Maybe<Array<Maybe<AggregationOption>>>;
  /** The relative position of the attribute in a layered navigation block. */
  position?: Maybe<Scalars['Int']['output']>;
};
// ...
export type ComplexTextValue = {
  __typename?: 'ComplexTextValue';
  /** Text that can contain HTML tags. */
  html: Scalars['String']['output'];
};
// ...
export enum CurrencyEnum {
  Aed = 'AED',
  Afn = 'AFN',
  All = 'ALL',
  Amd = 'AMD',
  Ang = 'ANG',
  Aoa = 'AOA',
  Ars = 'ARS',
  Aud = 'AUD',
  Awg = 'AWG',
  Azm = 'AZM',
  Azn = 'AZN',
  Bam = 'BAM',
  Bbd = 'BBD',
  Bdt = 'BDT',
  Bgn = 'BGN',
  Bhd = 'BHD',
  Bif = 'BIF',
  Bmd = 'BMD',
  Bnd = 'BND',
  Bob = 'BOB',
  Brl = 'BRL',
  Bsd = 'BSD',
  Btn = 'BTN',
  Buk = 'BUK',
  Bwp = 'BWP',
  Byn = 'BYN',
  Bzd = 'BZD',
  Cad = 'CAD',
  Cdf = 'CDF',
  Che = 'CHE',
  Chf = 'CHF',
  Chw = 'CHW',
  Clp = 'CLP',
  Cny = 'CNY',
  Cop = 'COP',
  Crc = 'CRC',
  Cup = 'CUP',
  Cve = 'CVE',
  Czk = 'CZK',
  Djf = 'DJF',
  Dkk = 'DKK',
  Dop = 'DOP',
  Dzd = 'DZD',
  Eek = 'EEK',
  Egp = 'EGP',
  Ern = 'ERN',
  Etb = 'ETB',
  Eur = 'EUR',
  Fjd = 'FJD',
  Fkp = 'FKP',
  Gbp = 'GBP',
  Gek = 'GEK',
  Gel = 'GEL',
  Ghs = 'GHS',
  Gip = 'GIP',
  Gmd = 'GMD',
  Gnf = 'GNF',
  Gqe = 'GQE',
  Gtq = 'GTQ',
  Gyd = 'GYD',
  Hkd = 'HKD',
  Hnl = 'HNL',
  Hrk = 'HRK',
  Htg = 'HTG',
  Huf = 'HUF',
  Idr = 'IDR',
  Ils = 'ILS',
  Inr = 'INR',
  Iqd = 'IQD',
  Irr = 'IRR',
  Isk = 'ISK',
  Jmd = 'JMD',
  Jod = 'JOD',
  Jpy = 'JPY',
  Kes = 'KES',
  Kgs = 'KGS',
  Khr = 'KHR',
  Kmf = 'KMF',
  Kpw = 'KPW',
  Krw = 'KRW',
  Kwd = 'KWD',
  Kyd = 'KYD',
  Kzt = 'KZT',
  Lak = 'LAK',
  Lbp = 'LBP',
  Lkr = 'LKR',
  Lrd = 'LRD',
  Lsl = 'LSL',
  Lsm = 'LSM',
  Ltl = 'LTL',
  Lvl = 'LVL',
  Lyd = 'LYD',
  Mad = 'MAD',
  Mdl = 'MDL',
  Mga = 'MGA',
  Mkd = 'MKD',
  Mmk = 'MMK',
  Mnt = 'MNT',
  Mop = 'MOP',
  Mro = 'MRO',
  Mur = 'MUR',
  Mvr = 'MVR',
  Mwk = 'MWK',
  Mxn = 'MXN',
  Myr = 'MYR',
  Mzn = 'MZN',
  Nad = 'NAD',
  Ngn = 'NGN',
  Nic = 'NIC',
  Nok = 'NOK',
  Npr = 'NPR',
  Nzd = 'NZD',
  Omr = 'OMR',
  Pab = 'PAB',
  Pen = 'PEN',
  Pgk = 'PGK',
  Php = 'PHP',
  Pkr = 'PKR',
  Pln = 'PLN',
  Pyg = 'PYG',
  Qar = 'QAR',
  Rhd = 'RHD',
  Rol = 'ROL',
  Ron = 'RON',
  Rsd = 'RSD',
  Rub = 'RUB',
  Rwf = 'RWF',
  Sar = 'SAR',
  Sbd = 'SBD',
  Scr = 'SCR',
  Sdg = 'SDG',
  Sek = 'SEK',
  Sgd = 'SGD',
  Shp = 'SHP',
  Skk = 'SKK',
  Sll = 'SLL',
  Sos = 'SOS',
  Srd = 'SRD',
  Std = 'STD',
  Svc = 'SVC',
  Syp = 'SYP',
  Szl = 'SZL',
  Thb = 'THB',
  Tjs = 'TJS',
  Tmm = 'TMM',
  Tnd = 'TND',
  Top = 'TOP',
  Trl = 'TRL',
  Try = 'TRY',
  Ttd = 'TTD',
  Twd = 'TWD',
  Tzs = 'TZS',
  Uah = 'UAH',
  Ugx = 'UGX',
  Usd = 'USD',
  Uyu = 'UYU',
  Uzs = 'UZS',
  Veb = 'VEB',
  Vef = 'VEF',
  Vnd = 'VND',
  Vuv = 'VUV',
  Wst = 'WST',
  Xcd = 'XCD',
  Xof = 'XOF',
  Xpf = 'XPF',
  Yer = 'YER',
  Ytl = 'YTL',
  Zar = 'ZAR',
  Zmk = 'ZMK',
  Zwd = 'ZWD'
}
// ...
export type LayerFilter = {
  __typename?: 'LayerFilter';
  /**
   * An array of filter items.
   * @deprecated Use `Aggregation.options` instead.
   */
  filter_items?: Maybe<Array<Maybe<LayerFilterItemInterface>>>;
  /**
   * The count of filter items in filter group.
   * @deprecated Use `Aggregation.count` instead.
   */
  filter_items_count?: Maybe<Scalars['Int']['output']>;
  /**
   * The name of a layered navigation filter.
   * @deprecated Use `Aggregation.label` instead.
   */
  name?: Maybe<Scalars['String']['output']>;
  /**
   * The request variable name for a filter query.
   * @deprecated Use `Aggregation.attribute_code` instead.
   */
  request_var?: Maybe<Scalars['String']['output']>;
};
// ...
export type Money = {
  __typename?: 'Money';
  /** A three-letter currency code, such as USD or EUR. */
  currency?: Maybe<CurrencyEnum>;
  /** A number expressing a monetary value. */
  value?: Maybe<Scalars['Float']['output']>;
};
// ...
export type Price = {
  __typename?: 'Price';
  /**
   * An array that provides information about tax, weee, or weee_tax adjustments.
   * @deprecated Use `ProductPrice` instead.
   */
  adjustments?: Maybe<Array<Maybe<PriceAdjustment>>>;
  /**
   * The price of a product plus a three-letter currency code.
   * @deprecated Use `ProductPrice` instead.
   */
  amount?: Maybe<Money>;
};
// ...
export type ProductImage = MediaGalleryInterface & {
  __typename?: 'ProductImage';
  /** Indicates whether the image is hidden from view. */
  disabled?: Maybe<Scalars['Boolean']['output']>;
  /** The label of the product image or video. */
  label?: Maybe<Scalars['String']['output']>;
  /** The media item's position after it has been sorted. */
  position?: Maybe<Scalars['Int']['output']>;
  /** The URL of the product image or video. */
  url?: Maybe<Scalars['String']['output']>;
};
// ...
export type ProductPrices = {
  __typename?: 'ProductPrices';
  /**
   * The highest possible final price for all the options defined within a composite product. If you are specifying a price range, this would be the `to` value.
   * @deprecated Use `PriceRange.maximum_price` instead.
   */
  maximalPrice?: Maybe<Price>;
  /**
   * The lowest possible final price for all the options defined within a composite product. If you are specifying a price range, this would be the `from` value.
   * @deprecated Use `PriceRange.minimum_price` instead.
   */
  minimalPrice?: Maybe<Price>;
  /**
   * The base price of a product.
   * @deprecated Use `regular_price` from `PriceRange.minimum_price` or `PriceRange.maximum_price` instead.
   */
  regularPrice?: Maybe<Price>;
};
// ...
export type Products = {
  __typename?: 'Products';
  /** A bucket that contains the attribute code and label for each filterable option. */
  aggregations?: Maybe<Array<Maybe<Aggregation>>>;
  /**
   * Layered navigation filters array.
   * @deprecated Use `aggregations` instead.
   */
  filters?: Maybe<Array<Maybe<LayerFilter>>>;
  /** An array of products that match the specified search criteria. */
  items?: Maybe<Array<Maybe<ProductInterface>>>;
  /** An object that includes the page_info and currentPage values specified in the query. */
  page_info?: Maybe<SearchResultPageInfo>;
  /** An object that includes the default sort field and all available sort fields. */
  sort_fields?: Maybe<SortFields>;
  /** An array of search suggestions for case when search query have no results. */
  suggestions?: Maybe<Array<Maybe<SearchSuggestion>>>;
  /** The number of products that are marked as visible. By default, in complex products, parent products are visible, but their child products are not. */
  total_count?: Maybe<Scalars['Int']['output']>;
};
// ...
export type SearchResultPageInfo = {
  __typename?: 'SearchResultPageInfo';
  /** The specific page to return. */
  current_page?: Maybe<Scalars['Int']['output']>;
  /** The maximum number of items to return per page of results. */
  page_size?: Maybe<Scalars['Int']['output']>;
  /** The total number of pages in the response. */
  total_pages?: Maybe<Scalars['Int']['output']>;
};

/** A string that contains search suggestion */
export type SearchSuggestion = {
  __typename?: 'SearchSuggestion';
  /** The search suggestion of existing product. */
  search: Scalars['String']['output'];
};
// ...
export type ProductsQuery = { __typename?: 'Query', products?: { __typename?: 'Products', total_count?: number | null, items?: Array<{ __typename?: 'BundleProduct', manufacturer?: number | null, meta_description?: string | null, meta_title?: string | null, meta_keyword?: string | null, name?: string | null, sku?: string | null, description?: { __typename?: 'ComplexTextValue', html: string } | null, image?: { __typename?: 'ProductImage', url?: string | null } | null, price?: { __typename?: 'ProductPrices', regularPrice?: { __typename?: 'Price', amount?: { __typename?: 'Money', currency?: CurrencyEnum | null, value?: number | null } | null } | null } | null, thumbnail?: { __typename?: 'ProductImage', url?: string | null } | null } | { __typename?: 'ConfigurableProduct', manufacturer?: number | null, meta_description?: string | null, meta_title?: string | null, meta_keyword?: string | null, name?: string | null, sku?: string | null, description?: { __typename?: 'ComplexTextValue', html: string } | null, image?: { __typename?: 'ProductImage', url?: string | null } | null, price?: { __typename?: 'ProductPrices', regularPrice?: { __typename?: 'Price', amount?: { __typename?: 'Money', currency?: CurrencyEnum | null, value?: number | null } | null } | null } | null, thumbnail?: { __typename?: 'ProductImage', url?: string | null } | null } | { __typename?: 'DownloadableProduct', manufacturer?: number | null, meta_description?: string | null, meta_title?: string | null, meta_keyword?: string | null, name?: string | null, sku?: string | null, description?: { __typename?: 'ComplexTextValue', html: string } | null, image?: { __typename?: 'ProductImage', url?: string | null } | null, price?: { __typename?: 'ProductPrices', regularPrice?: { __typename?: 'Price', amount?: { __typename?: 'Money', currency?: CurrencyEnum | null, value?: number | null } | null } | null } | null, thumbnail?: { __typename?: 'ProductImage', url?: string | null } | null } | { __typename?: 'GiftCardProduct', manufacturer?: number | null, meta_description?: string | null, meta_title?: string | null, meta_keyword?: string | null, name?: string | null, sku?: string | null, description?: { __typename?: 'ComplexTextValue', html: string } | null, image?: { __typename?: 'ProductImage', url?: string | null } | null, price?: { __typename?: 'ProductPrices', regularPrice?: { __typename?: 'Price', amount?: { __typename?: 'Money', currency?: CurrencyEnum | null, value?: number | null } | null } | null } | null, thumbnail?: { __typename?: 'ProductImage', url?: string | null } | null } | { __typename?: 'GroupedProduct', manufacturer?: number | null, meta_description?: string | null, meta_title?: string | null, meta_keyword?: string | null, name?: string | null, sku?: string | null, description?: { __typename?: 'ComplexTextValue', html: string } | null, image?: { __typename?: 'ProductImage', url?: string | null } | null, price?: { __typename?: 'ProductPrices', regularPrice?: { __typename?: 'Price', amount?: { __typename?: 'Money', currency?: CurrencyEnum | null, value?: number | null } | null } | null } | null, thumbnail?: { __typename?: 'ProductImage', url?: string | null } | null } | { __typename?: 'SimpleProduct', manufacturer?: number | null, meta_description?: string | null, meta_title?: string | null, meta_keyword?: string | null, name?: string | null, sku?: string | null, description?: { __typename?: 'ComplexTextValue', html: string } | null, image?: { __typename?: 'ProductImage', url?: string | null } | null, price?: { __typename?: 'ProductPrices', regularPrice?: { __typename?: 'Price', amount?: { __typename?: 'Money', currency?: CurrencyEnum | null, value?: number | null } | null } | null } | null, thumbnail?: { __typename?: 'ProductImage', url?: string | null } | null } | { __typename?: 'VirtualProduct', manufacturer?: number | null, meta_description?: string | null, meta_title?: string | null, meta_keyword?: string | null, name?: string | null, sku?: string | null, description?: { __typename?: 'ComplexTextValue', html: string } | null, image?: { __typename?: 'ProductImage', url?: string | null } | null, price?: { __typename?: 'ProductPrices', regularPrice?: { __typename?: 'Price', amount?: { __typename?: 'Money', currency?: CurrencyEnum | null, value?: number | null } | null } | null } | null, thumbnail?: { __typename?: 'ProductImage', url?: string | null } | null } | null> | null } | null };